+++ src/services/estimateService.ts
@@ -870,27 +870,6 @@ async function findMovingCompleteList(
   }
 }
 
-// 이사 완료 정보 수정 API
-async function updateDatabase() {
-  const today = todayUTC();
-  const estimateList = await estimateRepository.findManyData({
-    where: {
-      status: 'ACCEPTED',
-      isMovingComplete: false,
-      MovingInfo: { movingDate: { lt: today } },
-    },
-    select: estimateSelect,
-  });
-
-  await Promise.all(
-    estimateList.map(async (estimate) => {
-      await estimateRepository.updateData({
-        where: { id: estimate.id },
-        data: { isMovingComplete: true },
-      });
-    })
-  );
-}
 
 export default {
   findReceivedEstimateList,
@@ -901,5 +880,4 @@ export default {
   createEstimate,
   findEstimateDetail,
   findMovingCompleteList,
-  updateDatabase,
 };
